Update it_IT.po
Hi, I stumbled on your project and found it cool, so I took some time to edit/finalise the IT.po
I'm not sure if you want to use a formal or more "personal" language as it is not always consistent in the translations. I have also corrected some typos and tried to keep the same format for the repeatedly used words. Hope this helps. For your information, I'm native Italian.